Tentative processFood function

+++ Cooking with Cain/Assets/Scripts/Food.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Food : MonoBehaviour {
-    public int multiplier;
+    public float multiplier;
     public string foodName;
     public string[] attributes;
 }
